+++ plugins/org.preesm.ui.scenario/src/org/preesm/ui/scenario/editor/timings/TimingImporter.java
@@ -37,6 +37,7 @@ package org.preesm.ui.scenario.editor.timings;
 import org.preesm.model.scenario.Scenario;
 import org.preesm.model.scenario.serialize.CsvTimingParser;
 import org.preesm.model.scenario.serialize.ExcelTimingParser;
+import org.preesm.model.scenario.serialize.PapifyOutputTimingParser;
 import org.preesm.model.scenario.serialize.PapifyTimingParser;
 import org.preesm.model.slam.Design;
 
@@ -54,6 +55,7 @@ public class TimingImporter {
       final ExcelTimingParser excelParser = new ExcelTimingParser(currentScenario);
       final CsvTimingParser csvParser = new CsvTimingParser(currentScenario);
       final PapifyTimingParser papifyParser = new PapifyTimingParser(currentScenario);
+      final PapifyOutputTimingParser papifyOutputParser = new PapifyOutputTimingParser(currentScenario);
 
       try {
         final String[] fileExt = excelFileURL.split("\\.");
@@ -69,6 +71,9 @@ public class TimingImporter {
             papifyParser.parse(excelFileURL, design.getOperatorComponents());
             break;
           default:
+        }
+        if (excelFileURL.endsWith("papify-output")) {
+          papifyOutputParser.parse(excelFileURL, design.getOperatorComponents());
         }
       } catch (final Exception e) {
         e.printStackTrace();